Update validations and tests.
